+++ src/main/java/com/example/springdatabasicdemo/repositories/TruckRepository.java
@@ -1,14 +1,10 @@
 package com.example.springdatabasicdemo.repositories;
 
 import com.example.springdatabasicdemo.models.Truck;
+import org.hibernate.mapping.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface TruckRepository extends JpaRepository<Truck,Long>{
-
+public interface TruckRepository extends JpaRepository <Truck, Integer> {
 }
-
-
